Adds _WPSRequest class.
Replaces owslib's HTTP backend, exposing sessions management. Session
id's stored in cookies are now persistent during a session and will
eventually be persistent between sessions. CSRF tokens are now handled
without monkey patching, if they are present the HTTP POST request is
augmented to x-www-form-urlencoded which is written to the body rather
than the raw xml when the token is not present.